idc: move code from a header to a .c file
src/arch/xtensa/smp/include/arch/idc.h has way too much code for a
header, it is included by platform/idc.h on various platforms, and
that header in turn is included by bultiple .c files, which builds
the code multiple times into the firmware. Move it to a C file.
